네트워크 설정 및 Connect 기능 추가
connectex를 사용하기 위해서는 소켓을 먼저 iocp등록 한 후 bind를 호출(이때 sockaddr_in 구조체의 family를 제외하고 모두 0으로 초기화 해준다.)하고 connectex를 실행해야한다.

+++ BACore/NetMessage.h
@@ -10,9 +10,6 @@ struct Data
 };
 #pragma pack(pop)
 
-/*
-* 버퍼 타입으로 넘어오는 클래스는 GetSize()함수가 구현되어 있어야한다.
-*/
 class NetMessage
 {
 	friend class BASocket;
@@ -33,7 +30,7 @@ public:
 	template<typename T>
 	T* GetBuffer() { return reinterpret_cast<T>(_data._array); }
 
-	WORD GetSize() { return _size - sizeof(_size); }
+	ULONG GetSize() { return _size - sizeof(_size); }
 
 	template<typename T>
 	void SetSize() { _size = sizeof(_size) + reinterpret_cast<T>(_data._array).GetSize(); }
